backend/requirements.txt
# Intel Platform MVP - Python Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
